etl/requirements.txt
numpy==2.3.5
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
selenium==4.38.0
snowflake==1.9.0
